Bumped dependencies to their latest version (#82)
* :arrow_up: bumped dependencies

* :rotating_light: made analyzer happy

* :fire: removed unused provider from example app

+++ test/defaults/providers/rest/rest_service_test.dart
@@ -367,7 +367,7 @@ void main() {
   });
 
   tearDown(() {
-    file.delete();
+    if (file.existsSync()) file.delete();
   });
 }
